requirements.txt
db-dtypes==1.1.1
google-api-core==2.11.0
google-auth==2.17.1
google-cloud-bigquery==3.9.0
google-cloud-bigquery-datatransfer==3.11.1
google-cloud-core==2.3.2